requirements.txt
numpy
pandas
matplotlib
scipy
scikit-image
imagecodecs[all]
biopython
pillow
ttkbootstrap